+++ lib/xdr/double_float.ex
@@ -1,36 +1,32 @@
 defmodule XDR.DoubleFloat do
-  @behaviour XDR.Declaration
   @moduledoc """
-  This module is in charge of process the Double-Precision Floating-Point types based on the RFC4506 XDR Standard
+  This module manages the `Double-Precision Floating-Point` type based on the RFC4506 XDR Standard.
   """
 
-  defstruct float: nil
+  @behaviour XDR.Declaration
 
-  @typedoc """
-  Every double float structure has a float which represent the Double-Precision Floating-Point value which you try to encode
-  """
-  @type t :: %XDR.DoubleFloat{float: integer | float | binary}
+  alias XDR.Error.DoubleFloat, as: DoubleFloatError
 
-  alias XDR.Error.DoubleFloat
-
-  @doc """
-  this function provides an easy way to create an XDR.DoubleFloat type
-
-  returns a XDR.DoubleFloat struct with the value received as parameter
-  """
-  @spec new(float :: float | integer | binary) :: t()
-  def new(float), do: %XDR.DoubleFloat{float: float}
+  defstruct [:float]
 
   defguard valid_float?(value) when is_float(value) or is_integer(value)
 
+  @typedoc """
+  `XDR.DoubleFloat` struct type specification.
+  """
+  @type t :: %XDR.DoubleFloat{float: integer | float | binary}
+
+  @doc """
+  Create a new `XDR.DoubleFloat` structure from the `float` passed.
+  """
+  @spec new(float :: float | integer | binary) :: t
+  def new(float), do: %XDR.DoubleFloat{float: float}
+
   @impl XDR.Declaration
   @doc """
-  this function is in charge of encoding a Double Floating Point value into an XDR if the parameters are wrong an error will be raised,
-  it receives an XDR.DoubleFloat structure which contains the Double-Precision Floating-Point to encode
-
-  Returns a tuple with the XDR resulted from encoding the Double Floating Point value
+  Encode a `XDR.DoubleFloat` structure into a XDR format.
   """
-  @spec encode_xdr(t()) :: {:ok, binary()} | {:error, :not_number}
+  @spec encode_xdr(double_float :: t) :: {:ok, binary()} | {:error, :not_number}
   def encode_xdr(%XDR.DoubleFloat{float: float}) when not valid_float?(float),
     do: {:error, :not_number}
 
@@ -38,33 +34,27 @@ defmodule XDR.DoubleFloat do
 
   @impl XDR.Declaration
   @doc """
-  this function is in charge of encoding a Double Floating Point value into an XDR if the parameters are wrong an error will be raised,
-  it receives an XDR.DoubleFloat structure which contains the Double-Precision Floating-Point to encode
-
-  Returns the XDR resulted from encoding the Double Floating Point value
+  Encode a `XDR.DoubleFloat` structure into a XDR format.
+  If the `double_float` is not valid, an exception is raised.
   """
-  @spec encode_xdr!(t()) :: binary()
-  def encode_xdr!(float) do
-    case encode_xdr(float) do
+  @spec encode_xdr!(double_float :: t) :: binary()
+  def encode_xdr!(double_float) do
+    case encode_xdr(double_float) do
       {:ok, binary} -> binary
-      {:error, reason} -> raise(DoubleFloat, reason)
+      {:error, reason} -> raise(DoubleFloatError, reason)
     end
   end
 
   @impl XDR.Declaration
   @doc """
-  This function is in charge of decode the XDR value which represents an Double Floating Point value if the parameters are wrong
-  an error will be raised, it receives an XDR.DoubleFloat structure which contanis the binary to decode
-
-  Returns a tuple with the Double Floating Point resulted from decode the XDR value and its remaining bits
+  Decode the Double-Precision Floating-Point in XDR format to a `XDR.DoubleFloat` structure.
   """
-  @spec decode_xdr(bytes :: binary, opts :: any) :: {:ok, {t(), binary()}} | {:error, :not_binary}
-  def decode_xdr(bytes, opts \\ nil)
+  @spec decode_xdr(bytes :: binary, double_float :: t) ::
+          {:ok, {t, binary()}} | {:error, :not_binary}
+  def decode_xdr(bytes, double_float \\ nil)
+  def decode_xdr(bytes, _double_float) when not is_binary(bytes), do: {:error, :not_binary}
 
-  def decode_xdr(bytes, _opts) when not is_binary(bytes),
-    do: {:error, :not_binary}
-
-  def decode_xdr(bytes, _opts) do
+  def decode_xdr(bytes, _double_float) do
     <<float::big-signed-float-size(64), rest::binary>> = bytes
 
     decoded_float = new(float)
@@ -74,18 +64,16 @@ defmodule XDR.DoubleFloat do
 
   @impl XDR.Declaration
   @doc """
-  This function is in charge of decode the XDR value which represents an Double Floating Point value if the parameters are wrong
-  an error will be raised, it receives an XDR.DoubleFloat structure which contanis the binary to decode
-
-  Returns the Double Floating Point resulted from decode the XDR value and its remaining bits
+  Decode the Double-Precision Floating-Point in XDR format to a `XDR.DoubleFloat` structure.
+  If the binaries are not valid, an exception is raised.
   """
-  @spec decode_xdr!(bytes :: binary, opts :: any) :: {t(), binary()}
-  def decode_xdr!(bytes, opts \\ nil)
+  @spec decode_xdr!(bytes :: binary, double_float :: t) :: {t, binary()}
+  def decode_xdr!(bytes, double_float \\ nil)
 
-  def decode_xdr!(bytes, opts) do
-    case decode_xdr(bytes, opts) do
+  def decode_xdr!(bytes, double_float) do
+    case decode_xdr(bytes, double_float) do
       {:ok, result} -> result
-      {:error, reason} -> raise(DoubleFloat, reason)
+      {:error, reason} -> raise(DoubleFloatError, reason)
     end
   end
 end