+++ lib/xdr/enum.ex
@@ -1,27 +1,28 @@
 defmodule XDR.Enum do
-  @behaviour XDR.Declaration
   @moduledoc """
-  This module is in charge to process the Enum type based on the RFC4506 XDR Standard
+  This module manages the `Enumeration` type based on the RFC4506 XDR Standard.
   """
 
-  alias XDR.Error.Enum, as: EnumErr
+  @behaviour XDR.Declaration
 
-  defstruct declarations: nil, identifier: nil
+  alias XDR.Error.Enum, as: EnumError
+
+  defstruct [:declarations, :identifier]
 
   @typedoc """
-  Every enum structure has a declaration list which contains the keys and its representation value
+  `XDR.Enum` structure type specification.
   """
-  @type t :: %XDR.Enum{declarations: keyword, identifier: atom}
+  @type t :: %XDR.Enum{declarations: keyword(), identifier: atom()}
 
+  @doc """
+  Create a new `XDR.Enum` structure with the `declarations` and `identifier` passed.
+  """
   def new(declarations, identifier),
     do: %XDR.Enum{declarations: declarations, identifier: identifier}
 
   @impl XDR.Declaration
   @doc """
-  This function is in charge of encode a value inside of the enum structure based on the identifier received by parameter, it receives
-  a XDR.Enum type which contains the Enum and the identifier which you need to encode
-
-  Returns a tuple with the the XDR resulted from encode the value wich represents a key in the enum structure
+  Encode a `XDR.Enum` structure into a XDR format.
   """
   @spec encode_xdr(map) :: {:ok, binary} | {:error, :not_list | :not_an_atom | :invalid_key}
   def encode_xdr(%{declarations: declarations}) when not is_list(declarations),
@@ -31,88 +32,72 @@ defmodule XDR.Enum do
     do: {:error, :not_an_atom}
 
   def encode_xdr(%{declarations: declarations, identifier: identifier}) do
-    Keyword.has_key?(declarations, identifier)
-    |> encode_valid_data(declarations, identifier)
-  end
-
-  @impl XDR.Declaration
-  @doc """
-  This function is in charge of encode a value inside of the enum structure based on the identifier received by parameter, it receives
-  a XDR.Enum type which contains the Enum and the identifier which you need to encode
-
-  Returns the XDR resulted from encode the value wich represents a key in the enum structure
-  """
-  @spec encode_xdr!(enum :: map) :: binary
-  def encode_xdr!(enum) do
-    case encode_xdr(enum) do
-      {:ok, binary} -> binary
-      {:error, reason} -> raise(EnumErr, reason)
+    case Keyword.has_key?(declarations, identifier) do
+      true -> encode_valid_data(declarations, identifier)
+      false -> {:error, :invalid_key}
     end
   end
 
   @impl XDR.Declaration
   @doc """
-  This function is in charge of decode the XDR which represents a value inside an enum structure,it receives an XDR.Enum structure that
-  contains the identifier and the enum which it belongs
-
-  Returns a tuple with the key of the decoded enum and the remaining bytes if there are.
+  Encode a `XDR.Enum` structure into a XDR format.
+  If the `declarations` or `identifier` of the `enum` is not valid, an exception is raised.
   """
-  @spec decode_xdr(binary, map) ::
-          {:ok, {t(), binary}} | {:error, :not_binary | :not_list | :invalid_key}
-  def decode_xdr(bytes, %{}) when not is_binary(bytes),
-    do: {:error, :not_binary}
+  @spec encode_xdr!(enum :: t) :: binary
+  def encode_xdr!(enum) do
+    case encode_xdr(enum) do
+      {:ok, binary} -> binary
+      {:error, reason} -> raise(EnumError, reason)
+    end
+  end
+
+  @impl XDR.Declaration
+  @doc """
+  Decode the Enumeration in XDR format to a `XDR.Enum` structure.
+  """
+  @spec decode_xdr(bytes :: binary(), enum :: t) ::
+          {:ok, {t, binary}} | {:error, :not_binary | :not_list | :invalid_key}
+  def decode_xdr(bytes, %{}) when not is_binary(bytes), do: {:error, :not_binary}
 
   def decode_xdr(_bytes, %{declarations: declarations}) when not is_list(declarations),
     do: {:error, :not_list}
 
   def decode_xdr(bytes, %{declarations: declarations}) do
-    {value, rest} = XDR.Int.decode_xdr!(bytes)
+    {int_xdr, rest} = XDR.Int.decode_xdr!(bytes)
 
-    identifier = Enum.find(declarations, &Kernel.===(elem(&1, 1), value.datum)) |> get_response()
+    declarations
+    |> Enum.find(fn {_key, value} -> value === int_xdr.datum end)
+    |> decoded_enum(declarations, rest)
+  end
 
+  @spec decoded_enum(
+          identifier :: {atom(), any()} | nil,
+          declarations :: keyword(),
+          rest :: binary()
+        ) :: {:ok, {t, binary}} | {:error, :invalid_key}
+  defp decoded_enum(nil, _declarations, _rest), do: {:error, :invalid_key}
+
+  defp decoded_enum({identifier, _value}, declarations, rest) do
     decoded_enum = new(declarations, identifier)
-
-    Keyword.has_key?(declarations, identifier)
-    |> add_valid_key(decoded_enum, rest)
+    {:ok, {decoded_enum, rest}}
   end
 
   @impl XDR.Declaration
   @doc """
-  This function is in charge of decode the XDR which represents a value inside an enum structure,it receives an XDR.Enum structure that
-  contains the identifier and the enum which it belongs
-
-  Returns the key of the decoded enum and the remaining bytes if there are.
+  Decode the Enumeration in XDR format to a `XDR.Enum` structure.
+  If the binaries are not valid, an exception is raised.
   """
-  @spec decode_xdr!(binary, map) :: {t(), binary}
-  def decode_xdr!(bytes, struct) do
-    case decode_xdr(bytes, struct) do
+  @spec decode_xdr!(bytes :: binary(), enum :: t) :: {t, binary()}
+  def decode_xdr!(bytes, enum) do
+    case decode_xdr(bytes, enum) do
       {:ok, result} -> result
-      {:error, reason} -> raise(EnumErr, reason)
+      {:error, reason} -> raise(EnumError, reason)
     end
   end
 
-  @spec get_response({atom(), any}) :: atom()
-  defp get_response({identifier, _}), do: identifier
-
-  @spec get_response(nil) :: nil
-  defp get_response(nil), do: nil
-
-  @spec add_valid_key(boolean(), map(), rest :: binary()) ::
-          {:ok, {t, binary}} | {:error, :invalid_key}
-  defp add_valid_key(true, decoded_enum, rest), do: {:ok, {decoded_enum, rest}}
-
-  defp add_valid_key(false, _enum, _rest), do: {:error, :invalid_key}
-
-  @spec encode_valid_data(boolean(), declarations :: keyword(), identifier :: atom()) ::
-          {:ok, binary()}
-  defp encode_valid_data(true, declarations, identifier) do
-    binary =
-      declarations[identifier]
-      |> XDR.Int.new()
-      |> XDR.Int.encode_xdr!()
-
+  @spec encode_valid_data(declarations :: keyword(), identifier :: atom()) :: {:ok, binary()}
+  defp encode_valid_data(declarations, identifier) do
+    binary = declarations[identifier] |> XDR.Int.new() |> XDR.Int.encode_xdr!()
     {:ok, binary}
   end
-
-  defp encode_valid_data(false, _declarations, _identifier), do: {:error, :invalid_key}
 end