+++ lib/xdr/union.ex
@@ -1,26 +1,38 @@
 defmodule XDR.Union do
-  @behaviour XDR.Declaration
   @moduledoc """
-  this module is in charge of process the Discriminated Union types based on the RFC4506 XDR Standard
+  This module manages the `Discriminated Union` type based on the RFC4506 XDR Standard.
   """
 
-  defstruct discriminant: nil, arms: nil, value: nil
+  @behaviour XDR.Declaration
 
-  @type t :: %XDR.Union{discriminant: XDR.Enum | XDR.Int | XDR.UInt, arms: list}
+  alias XDR.Error.Union, as: UnionError
 
-  alias XDR.Error.Union
+  defstruct [:discriminant, :arms, :value]
 
+  @typedoc """
+  `XDR.Union` structure type specification.
+  """
+  @type t :: %XDR.Union{
+          discriminant: XDR.Enum.t() | XDR.Int.t() | XDR.UInt.t(),
+          arms: keyword() | map()
+        }
+
+  @doc """
+  Create a new `XDR.Union` structure with the `discriminant`, `arms` and `value` passed.
+  """
+  @spec new(
+          discriminant :: XDR.Enum.t() | XDR.Int.t() | XDR.UInt.t(),
+          arms :: keyword() | map(),
+          value :: any()
+        ) :: t
   def new(discriminant, arms, value \\ nil),
     do: %XDR.Union{discriminant: discriminant, arms: arms, value: value}
 
   @impl XDR.Declaration
   @doc """
-  this function is in charge of encoding the Discriminated Union types, it receives a struct which contains the discriminant
-  which is the principal value to make the Union, and the arms which are the possible values that can be taken by the discriminant
-
-  returns an ok tuple which contains the binary encoded of the union
+  Encode a `XDR.Union` structure into a XDR format.
   """
-  @spec encode_xdr(map()) :: {:ok, binary()} | {:error, :not_atom}
+  @spec encode_xdr(union :: t) :: {:ok, binary()} | {:error, :not_atom}
   def encode_xdr(%{discriminant: %{identifier: identifier}}) when not is_atom(identifier),
     do: {:error, :not_atom}
 
@@ -32,7 +44,7 @@ defmodule XDR.Union do
     discriminant_module = discriminant.__struct__
     encoded_discriminant = discriminant_module.encode_xdr!(discriminant)
 
-    encoded_arm = arms[identifier] |> encode_arm(value)
+    encoded_arm = identifier |> get_arm(arms) |> encode_arm(value)
 
     {:ok, encoded_discriminant <> encoded_arm}
   end
@@ -41,45 +53,29 @@ defmodule XDR.Union do
     discriminant_module = discriminant.__struct__
     encoded_discriminant = discriminant_module.encode_xdr!(discriminant)
 
-    encoded_arm = arms[discriminant.datum] |> encode_arm(value)
+    encoded_arm = discriminant.datum |> get_arm(arms) |> encode_arm(value)
 
     {:ok, encoded_discriminant <> encoded_arm}
   end
 
-  @spec encode_arm(arm :: struct() | module(), value :: any()) :: binary()
-  defp encode_arm(%_{} = arm, nil) do
-    arm_module = arm.__struct__
-    arm_module.encode_xdr!(arm)
-  end
-
-  defp encode_arm(arm, value) when is_atom(arm) do
-    arm.new(value) |> arm.encode_xdr!()
-  end
-
   @impl XDR.Declaration
   @doc """
-  this function is in charge of encoding the Discriminated Union types, it receives a struct which contains the discriminant
-  which is the principal value to make the Union, and the arms which are the possible values that can be taken by the discriminant
-
-  returns the binary encoded of the union
+  Encode a `XDR.Union` structure into a XDR format.
+  If the `union` is not valid, an exception is raised.
   """
-  @spec encode_xdr!(map()) :: binary()
+  @spec encode_xdr!(union :: t) :: binary()
   def encode_xdr!(union) do
     case encode_xdr(union) do
       {:ok, binary} -> binary
-      {:error, reason} -> raise(Union, reason)
+      {:error, reason} -> raise(UnionError, reason)
     end
   end
 
   @impl XDR.Declaration
   @doc """
-  this function is in charge of decoding the XDR which represents an Disxriminated Union type, it receives a struct which contains the discriminant
-  which is the principal value to make the Union, and the arms which are the possible values that can be taken by the discriminant, and the specific
-  struct
-
-  returns an ok tuple which contains the values of the union
+  Decode the Discriminated Union in XDR format to a `XDR.Union` structure.
   """
-  @spec decode_xdr(bytes :: binary(), union :: map()) ::
+  @spec decode_xdr(bytes :: binary(), union :: t | map()) ::
           {:ok, {any, binary()}} | {:error, :not_binary | :not_list}
   def decode_xdr(bytes, union) do
     decode_union_discriminant(bytes, union)
@@ -88,17 +84,14 @@ defmodule XDR.Union do
 
   @impl XDR.Declaration
   @doc """
-  this function is in charge of decoding the XDR which represents an Disxriminated Union type, it receives a struct which contains the discriminant
-  which is the principal value to make the Union, and the arms which are the possible values that can be taken by the discriminant, and the specific
-  struct
-
-  returns the values of the union
+  Decode the Discriminated Union in XDR format to a `XDR.Union` structure.
+  If the binaries are not valid, an exception is raised.
   """
-  @spec decode_xdr!(bytes :: binary(), union :: map()) :: {any, binary()}
+  @spec decode_xdr!(bytes :: binary(), union :: t | map()) :: {any, binary()}
   def decode_xdr!(bytes, union) do
     case decode_xdr(bytes, union) do
       {:ok, result} -> result
-      {:error, reason} -> raise(Union, reason)
+      {:error, reason} -> raise(UnionError, reason)
     end
   end
 
@@ -127,10 +120,18 @@ defmodule XDR.Union do
 
   defp decode_union_discriminant(bytes, %{discriminant: discriminant} = union) do
     discriminant_module = discriminant.__struct__
-
     {%{datum: datum}, rest} = discriminant_module.decode_xdr!(bytes)
-
     {%{union | discriminant: datum}, rest}
+  end
+
+  @spec encode_arm(arm :: struct() | module(), value :: any()) :: binary()
+  defp encode_arm(%_{} = arm, nil) do
+    arm_module = arm.__struct__
+    arm_module.encode_xdr!(arm)
+  end
+
+  defp encode_arm(arm, value) when is_atom(arm) do
+    arm.new(value) |> arm.encode_xdr!()
   end
 
   @spec decode_union_arm({:error, atom}) :: {:error, atom}
@@ -138,14 +139,21 @@ defmodule XDR.Union do
 
   @spec decode_union_arm({map(), binary}) :: {:ok, {{atom | integer, any}, binary}}
   defp decode_union_arm({%{discriminant: discriminant, arms: arms}, rest}) do
-    arm_module = arms[discriminant] |> get_arm_module()
-
+    arm_module = discriminant |> get_arm(arms) |> get_arm_module()
     {decoded_arm, rest} = arm_module.decode_xdr!(rest)
-
     {:ok, {{discriminant, decoded_arm}, rest}}
   end
 
   @spec get_arm_module(arm :: struct() | module()) :: module()
   defp get_arm_module(%_{} = arm), do: arm.__struct__
   defp get_arm_module(arm) when is_atom(arm), do: arm
+
+  @spec get_arm(identifier :: atom() | number(), arms :: keyword() | map()) ::
+          struct() | module() | nil
+  defp get_arm(identifier, arms) do
+    case arms[identifier] do
+      nil -> arms[:default]
+      arm -> arm
+    end
+  end
 end