+++ lib/xdr/variable_opaque.ex
@@ -1,36 +1,32 @@
 defmodule XDR.VariableOpaque do
-  @behaviour XDR.Declaration
   @moduledoc """
-  This module is in charge of process the Variable length opaque based on the RFC4506 XDR Standard
+  This module manages the `Variable-Length Opaque Data` type based on the RFC4506 XDR Standard.
   """
 
-  defstruct opaque: nil, max_size: nil
+  @behaviour XDR.Declaration
+
+  alias XDR.{FixedOpaque, UInt}
+  alias XDR.Error.VariableOpaque, as: VariableOpaqueError
+
+  defstruct [:opaque, :max_size]
 
   @typedoc """
-  Every double float structure has a float which represent the Double-Precision Floating-Point value which you try to encode
+  `XDR.VariableOpaque` structure type specification.
   """
   @type t :: %XDR.VariableOpaque{opaque: binary() | nil, max_size: integer()}
 
-  alias XDR.{FixedOpaque, UInt}
-  alias XDR.Error.VariableOpaque, as: VariableOpaqueErr
-
   @doc """
-  this function provides an easy way to create an XDR.VariableOpaque type
-
-  returns a XDR.VariableOpaque struct with the value received as parameter
+  Create a new `XDR.VariableOpaque` structure with the `opaque` and `max_size` passed.
   """
-  @spec new(opaque :: binary | nil, max_size :: integer) :: t()
+  @spec new(opaque :: binary | nil, max_size :: integer) :: t
   def new(opaque, max_size \\ 4_294_967_295)
   def new(opaque, max_size), do: %XDR.VariableOpaque{opaque: opaque, max_size: max_size}
 
   @impl XDR.Declaration
   @doc """
-  this function is in charge of encode the variable length opaque into an XDR,it receives an XDR.VariableOpaque stucture
-  which contains the needed data to encode the opaque
-
-  returns an :ok tuple with the resulted XDR
+  Encode a `XDR.VariableOpaque` structure into a XDR format.
   """
-  @spec encode_xdr(map()) ::
+  @spec encode_xdr(opaque :: t) ::
           {:ok, binary}
           | {:error,
              :not_binary
@@ -56,42 +52,29 @@ defmodule XDR.VariableOpaque do
 
   def encode_xdr(%{opaque: opaque}) do
     length = byte_size(opaque)
-
-    opaque_length =
-      length
-      |> UInt.new()
-      |> UInt.encode_xdr!()
-
-    fixed_opaque =
-      FixedOpaque.new(opaque, length)
-      |> FixedOpaque.encode_xdr!()
-
+    opaque_length = length |> UInt.new() |> UInt.encode_xdr!()
+    fixed_opaque = FixedOpaque.new(opaque, length) |> FixedOpaque.encode_xdr!()
     {:ok, opaque_length <> fixed_opaque}
   end
 
   @impl XDR.Declaration
   @doc """
-  this function is in charge of encode the variable length opaque into an XDR,it receives an XDR.VariableOpaque stucture
-  which contains the needed data to encode the opaque
-
-  returns the resulted XDR
+  Encode a `XDR.VariableOpaque` structure into a XDR format.
+  If the `opaque` is not valid, an exception is raised.
   """
-  @spec encode_xdr!(map()) :: binary
+  @spec encode_xdr!(opaque :: t) :: binary()
   def encode_xdr!(opaque) do
     case encode_xdr(opaque) do
       {:ok, binary} -> binary
-      {:error, reason} -> raise(VariableOpaqueErr, reason)
+      {:error, reason} -> raise(VariableOpaqueError, reason)
     end
   end
 
   @impl XDR.Declaration
   @doc """
-  this function is in charge of decode the XDR into a variable length opaque,it receives an XDR.VariableOpaque stucture
-  which contains the needed data to decode the opaque
-
-  returns an :ok tuple with the resulted binary
+  Decode the Variable-Length Opaque Data in XDR format to a `XDR.VariableOpaque` structure.
   """
-  @spec decode_xdr(bytes :: binary, opts :: map()) ::
+  @spec decode_xdr(bytes :: binary, opaque :: t) ::
           {:ok, {t, binary}}
           | {:error,
              :not_binary
@@ -100,9 +83,9 @@ defmodule XDR.VariableOpaque do
              | :exceed_upper_bound
              | :length_over_max
              | :length_over_rest}
-  def decode_xdr(bytes, opts \\ %{max_size: 4_294_967_295})
+  def decode_xdr(bytes, opaque \\ %{max_size: 4_294_967_295})
 
-  def decode_xdr(bytes, _opts) when not is_binary(bytes),
+  def decode_xdr(bytes, _opaque) when not is_binary(bytes),
     do: {:error, :not_binary}
 
   def decode_xdr(_bytes, %{max_size: max_size}) when not is_integer(max_size),
@@ -116,41 +99,34 @@ defmodule XDR.VariableOpaque do
 
   def decode_xdr(bytes, %{max_size: max_size}) do
     {uint, rest} = UInt.decode_xdr!(bytes)
-
-    get_decoded_value({uint.datum, rest}, max_size)
+    uint.datum |> get_decoded_value(rest, max_size)
   end
 
   @impl XDR.Declaration
   @doc """
-  this function is in charge of decode the XDR into a variable length opaque,it receives an XDR.VariableOpaque stucture
-  which contains the needed data to decode the opaque
-
-  returns an :ok tuple with the resulted binary
+  Decode the Variable-Length Opaque Data in XDR format to a `XDR.VariableOpaque` structure.
+  If the binaries are not valid, an exception is raised.
   """
-  @spec decode_xdr!(bytes :: binary, opts :: map()) :: {t, binary}
-  def decode_xdr!(bytes, struct \\ %{max_size: 4_294_967_295})
+  @spec decode_xdr!(bytes :: binary, opaque :: t) :: {t, binary}
+  def decode_xdr!(bytes, opaque \\ %{max_size: 4_294_967_295})
 
-  def decode_xdr!(bytes, struct) do
-    case decode_xdr(bytes, struct) do
+  def decode_xdr!(bytes, opaque) do
+    case decode_xdr(bytes, opaque) do
       {:ok, result} -> result
-      {:error, reason} -> raise(VariableOpaqueErr, reason)
+      {:error, reason} -> raise(VariableOpaqueError, reason)
     end
   end
 
-  @spec get_decoded_value({integer(), binary()}, max :: integer()) :: {:ok, {t, binary}}
-  defp get_decoded_value({length, _rest}, max) when length > max,
-    do: {:error, :length_over_max}
+  @spec get_decoded_value(length :: integer(), rest :: binary(), max :: integer()) ::
+          {:ok, {t, binary}}
+  defp get_decoded_value(length, _rest, max) when length > max, do: {:error, :length_over_max}
 
-  defp get_decoded_value({length, rest}, _max) when length > byte_size(rest),
+  defp get_decoded_value(length, rest, _max) when length > byte_size(rest),
     do: {:error, :length_over_rest}
 
-  defp get_decoded_value({length, rest}, max) do
+  defp get_decoded_value(length, rest, max) do
     {fixed_opaque, rest} = FixedOpaque.decode_xdr!(rest, %XDR.FixedOpaque{length: length})
-
-    decoded_variable_array =
-      fixed_opaque.opaque
-      |> new(max)
-
+    decoded_variable_array = fixed_opaque.opaque |> new(max)
     {:ok, {decoded_variable_array, rest}}
   end
 end