+++ test/xdr/quad_float_test.exs
@@ -1,4 +1,8 @@
 defmodule XDR.QuadFloatTest do
+  @moduledoc """
+  Tests for the `XDR.QuadFloat` module.
+  """
+
   use ExUnit.Case
 
   alias XDR.QuadFloat